feat! Add FollowRepository.cs and move relevant methods
