+++ src/Chirp.Infrastructure/Chirp.Infrastructure.csproj
@@ -7,7 +7,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <Folder Include="Chirp.Repositories\" />
       <Folder Include="Chirp.Services\" />
     </ItemGroup>
